Processing tests [part 2] (#47922)
* .

* draft it

* fix some

* maybe

* .

* chat template in qwen style

* draft sth for video sampling tests

* fix

* .

* fix last

* move around

* draft

* .

* renamed those

* fix repo deleted one line in all files

* fix

* the rest

* merge this

* typo

* when will this end?

* fix?

* save coz now i will break stuff

* rename it all, except for xx_token_id

* missed these renames

* fix last

* revert

* comments

* less small private helpers

* more comments

* last failing tests, and i stop adding more stuff

* typo

* forgot about this one

* why i cannot get green CI

* move around to skip more tests

* few more

* fix

* new model

* here it is, opened PR on the hub

* forgot modular...

+++ src/transformers/models/glm46v/modular_glm46v.py
@@ -12,7 +12,6 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-
 import numpy as np
 from huggingface_hub.dataclasses import strict
 
@@ -179,7 +178,7 @@ class Glm46VVideoProcessor(Glm4vVideoProcessor):
         if len(uniq) & 1:
             uniq.append(uniq[-1])
 
-        return np.array(uniq)
+        return np.array(uniq, dtype=int)
 
 
 __all__ = [
