+++ src/transformers/models/glm5_next/modular_glm5_next.py
@@ -2187,7 +2187,7 @@ class Glm5NextVideoProcessor(GlmgaVideoProcessor):
         if len(uniq) & 1:
             uniq.append(uniq[-1])
 
-        return np.array(uniq)
+        return np.array(uniq, dtype=int)
 
     def _preprocess(
         self,
